Homework Mostly Done
I'm really late on getting this in so I'm just going to turn in what I
have. Saved states doesn't work  when the activity is destroyed and
recreated. If the orientation is not changed. everything works well. I
am definitely going to try and make a question class next time instead
of arraylists....